update calculate_missing_data to pass number of expected trials
Added a new parameter total_trials_expected to the clean_missing_data function, allowing users to specify the total number of trials expected for each participant when calculating the proportion of missing trials. If this parameter is left as NULL, the proportion is calculated based on the actual number of trials in the data (as before).